feat: add the slicket-store crate and the PostgreSQL schema
Adds slicket-store, which stores Slicket's data in PostgreSQL through
sqlx. Its init migration creates the schema, including a tenant table
and the ticket statuses that every ticket shares. A CHECK constraint
rejects a ticket whose closed_at comes before its created_at. Indexes
cover the foreign keys on person and ticket. DbConfig reads the
database URL from DATABASE_URL, and init_database runs the embedded
migrations against a connection pool.

slicket-core now uses signed database key types, and its Entity marker
types derive Ord. Removes the slicket-showcase crate.

Adds a Docker Compose file for the local PostgreSQL database. ADRs 0002
and 0003 record the choice of PostgreSQL and sqlx. The README explains
the dev and main branches and describes the tenant.

+++ crates/slicket-core/src/ticket.rs
@@ -116,29 +116,30 @@ impl TicketCore {
 /// The `Entity<Ticket>` is the ticket's index in the ECS, while the `TicketKey` is the ticket's key
 /// in the database.
 ///
-/// Only code inside `slicket-core` can create a `TicketKey`. The number is a `u32`, which allows
-/// 4,294,967,296 tickets.
+/// Only code inside `slicket-core` can create a `TicketKey`. The number is an `i32`. This is a
+/// decision: the database stores the key as a Postgres `INTEGER`, which is signed, and numbers keys
+/// from 1 upwards. An `i32` therefore allows 2,147,483,647 tickets.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct TicketKey(pub(crate) u32);
+pub struct TicketKey(pub(crate) i32);
 
 impl TicketKey {
     /// Returns the number that identifies this ticket.
-    pub fn get(self) -> u32 {
+    pub fn get(self) -> i32 {
         self.0
     }
 }
 
 /// A `StatusKey` is the key that identifies a ticket status in the database.
 ///
-/// Only code inside `slicket-core` can create a `StatusKey`. The number is a `u16`, which allows
-/// 65,536 statuses. This is a decision: a `u8` allows only 256, and a `u32` allows far more
-/// statuses than any organisation will define.
+/// Only code inside `slicket-core` can create a `StatusKey`. The number is an `i32`. This is a
+/// decision: the database stores the key as a Postgres `INTEGER`, which is signed, and numbers keys
+/// from 1 upwards. An `i32` therefore allows 2,147,483,647 statuses.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct StatusKey(pub(crate) u16);
+pub struct StatusKey(pub(crate) i32);
 
 impl StatusKey {
     /// Returns the number that identifies this status.
-    pub fn get(self) -> u16 {
+    pub fn get(self) -> i32 {
         self.0
     }
 }
